pyproject.toml
[project]
name = "email-digest"
version = "0.1.0"
description = "Automated daily email digest with Claude-powered analysis"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.50",
    "openai>=1.0",
    "yt-dlp>=2024.1",
    "youtube-transcript-api>=0.6",
    "requests>=2.32",
    "readability-lxml>=0.8",
    "playwright>=1.48",
    "feedparser>=6.0",
    "resend>=2.0",
    "jinja2>=3.1",
    "premailer>=3.10",
    "pillow>=10.0",
    "pyyaml>=6.0",
    "lxml>=5.0",
    "aiohttp>=3.9",
    "markdown>=3.6",
]

[project.scripts]
email-digest = "email_digest.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/email_digest"]

[dependency-groups]
dev = [
    "ruff>=0.8",
]
